requirements.txt
torch
transformers
datasets
peft
bitsandbytes
trl
accelerate
scikit-learn
fastapi
uvicorn
groq
python-dotenv